Caching is not working properly when the multi language support is enabled.
The CI caching process uses the requested URL as a base to generate the
hash of the file that stores the cache, but the information contained in
CI URI Class doesn't contain any reference to the requested language
(it's already parsed), always showing the first page stored.

To avoid this we will add the "LANGUAGE" constant, generated in the
"Language" hook, at the time of writing and displaying the cache,
overwriting the CI Output Class.